Automated image analysis of ciliated cells
Script for the automated image analysis we used to determine overlap between candidate proteins and a multiplex immunohistochemistry panel in ImageJ Fiji. It is a simple and straight forward method, using thresholding (Multi otsu) to segment each marker and generate regions of interest (ROIs) to measure pixel values of the candidate protein, telling us with which panel marker the protein overlaps with the most.